password.lunar: fix double password prompt
Sometimes it's asking for password twice--once in the dialog, and once
again when the password command is actually run.

Turns out there's a flag for the passwd command to tell it to accept
input from standard input even if that isn't the current controlling
tty.

Also I threw in an extra EOL in the printf to make sure the enter key
gets pressed both times.